+++ opentelemetry-instrumentation/src/opentelemetry/instrumentation/bootstrap_gen.py
@@ -116,6 +116,10 @@ libraries = {
         "library": "redis >= 2.6",
         "instrumentation": "opentelemetry-instrumentation-redis==0.31b0",
     },
+    "remoulade": {
+        "library": "remoulade >= 0.50",
+        "instrumentation": "opentelemetry-instrumentation-remoulade==0.31b0",
+    },
     "requests": {
         "library": "requests ~= 2.0",
         "instrumentation": "opentelemetry-instrumentation-requests==0.31b0",